pyproject.toml
[tool.poetry]
name = "pythonistas-website"
version = "0.1.0"
description = ""
authors = ["pythoninthegrass <4097471+pythoninthegrass@users.noreply.github.com>"]
license = "Unlicense"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-fasthtml = "^0.10.3"

[tool.poetry.group.dev.dependencies]
uv = "^0.5.13"
ruff = "^0.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
